feat: create composables for two factor fetch, used it on profile page
