+++ src/composables/api/auth/index.ts
@@ -1,3 +1,4 @@
 import { useLogin } from "./use-login";
+import { useTwoFactor } from "./use-two-factor";
 
-export { useLogin }
+export { useLogin, useTwoFactor }
